tech(api): migrates script to cleanup challenge translations to pg only
